requirements.txt
flask
sentence-transformers
pandas
scikit-learn
gunicorn
transformers
accelerate
peft
optimum
faiss-cpu
fuzzywuzzy